pyproject.toml
[project]
name = "auto-fingerprint"
version = "0.1.0"
description = ""
authors = [
    {name = "Armin Sabouri",email = "me@arminsabouri.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai (==0.28)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "tree-sitter (>=0.24.0,<0.25.0)",
    "tree-sitter-python (>=0.23.6,<0.24.0)"
]

[tool.poetry]
packages = [{include = "auto_fingerprint", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
